Some refactoring and reordering
