+++ src/enums/vcs.rs
@@ -1,3 +1,4 @@
+/// Supported VCS
 #[derive(Debug, PartialEq)]
 pub enum VersionControlSystem {
     Git,
